chore(web): add useInfiniteScroll & list refactor
